Use progressive enhancement instead of graceful degradation for mobile support
The default.css for duck's theme is the same content as the original theme,     
with some attributes removed or tweaked slightly. The original theme CSS          
overrides almost all of this file leaving the desktop style unchanged. Once the 
mobile CSS is more complete and the common CSS is identified, it can be removed 
from the desktop CSS.

The mobile dropdown menu was obtained from http://astuteo.com/mobilemenu and as
per the license on the page:
"... it’s entirely free for you to download and use, modify for your own
applications, or otherwise make millions off of."