Copy env's pybabel into base dir to fix string extraction
The jinja2 extracter needs to know about our custom HighlightExtension (used on
pages such as get-involved/develop/applications), which is not in the env's
python path. Copying pybabel into the base dir gives it access to our modules.
This is only required for string extraction; other scripts use pybabel from the
env directly.

If the env's pybabel version is updated, this copy must also be updated.