scripts: add optional support for "timeout"; allow site-updater to be called directly
Sometimes the mtn syncs hang until aborted. Well, my syncing with welterde and
echelon would hang periodically until I added the same timeout support I'm
adding here. In short, if Montone's action has not completed within 20 minutes,
forcably kill the process. /usr/bin/timeout is a part of coreutils so most
(all?) Linux distributions should have it but I don't require timeout to exist.
If "timeout" can be found, it'll be used.

Also, set site-updater.sh to change to the directory in which the script is
located. This way one can give an absolute path in a cronjob.

Both of these changes should be low risk.