donate: change bitcoin url to https to avoid a redirect; host flattr image ourselves
RE: flattr: Since we force https on the internet facing site, with the http
link there was a warning from the browser about 'mixed mode content' . Using
https://api.flattr.com/button/button-static-50x60.png caused the image to
disappear if one only accepts strong ciphers in the browser
(e.g.  security.tls.version.min set to 2 in Firefox). It would look like
we were doing something wrong even though it's not us, so that's not a viable option.

Looking over https://flattr.com/terms it seems that there's no problem with our
hosting it ourselves. It's also a tiny image besides.

Now to email flattr to strengthen the crypto on their site...